isaac_ros-dev/src/web_control_server/requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.33.0  # Optional, for better concurrency
rclpy>=0.10.0
geometry-msgs>=0.12.0
